Update PR to changes introduced to the webInspector
